chore(rules): split __init__.py
