+++ mergify_engine/rules/__init__.py
@@ -17,7 +17,6 @@ import dataclasses
 import functools
 import itertools
 import operator
-import textwrap
 import typing
 
 import daiquiri
@@ -27,367 +26,16 @@ import yaml
 from mergify_engine import actions
 from mergify_engine import context
 from mergify_engine import github_types
-from mergify_engine.clients import http
-from mergify_engine.rules import filter
+from mergify_engine.rules import conditions
 from mergify_engine.rules import live_resolvers
-from mergify_engine.rules import parser
 from mergify_engine.rules import types
 
 
 LOG = daiquiri.getLogger(__name__)
 
 
-# This helps mypy breaking the recursive definition
-FakeTreeT = typing.Dict[str, typing.Any]
-
-
-@dataclasses.dataclass
-class RuleCondition:
-    """This describe a leaf of the `conditions:` tree, eg:
-
-    label=foobar
-    -merged
-    """
-
-    condition: typing.Union[str, FakeTreeT]
-    description: typing.Optional[str] = None
-    partial_filter: filter.Filter[bool] = dataclasses.field(init=False)
-    match: bool = dataclasses.field(init=False, default=False)
-    _used: bool = dataclasses.field(init=False, default=False)
-    evaluation_error: typing.Optional[str] = dataclasses.field(init=False, default=None)
-
-    def __post_init__(self) -> None:
-        self.update(self.condition)
-
-    def update(self, condition_raw: typing.Union[str, FakeTreeT]) -> None:
-        self.condition = condition_raw
-        try:
-            if isinstance(condition_raw, str):
-                condition = parser.search.parseString(condition_raw, parseAll=True)[0]
-            else:
-                condition = condition_raw
-            self.partial_filter = filter.BinaryFilter(
-                typing.cast(filter.TreeT, condition)
-            )
-        except (parser.pyparsing.ParseException, filter.InvalidQuery) as e:
-            raise voluptuous.Invalid(
-                message=f"Invalid condition '{condition_raw}'. {str(e)}",
-                error_message=str(e),
-            )
-
-    def update_attribute_name(self, new_name: str) -> None:
-        tree = typing.cast(filter.TreeT, self.partial_filter.tree)
-        negate = "-" in tree
-        tree = tree.get("-", tree)
-        operator = list(tree.keys())[0]
-        name, value = list(tree.values())[0]
-        if name.startswith(filter.Filter.LENGTH_OPERATOR):
-            new_name = f"{filter.Filter.LENGTH_OPERATOR}{new_name}"
-
-        new_tree: FakeTreeT = {operator: (new_name, value)}
-        if negate:
-            new_tree = {"-": new_tree}
-        self.update(new_tree)
-
-    def __str__(self) -> str:
-        if isinstance(self.condition, str):
-            return self.condition
-        else:
-            return str(self.partial_filter)
-
-    def copy(self) -> "RuleCondition":
-        rc = RuleCondition(self.condition, self.description)
-        rc.partial_filter.value_expanders = self.partial_filter.value_expanders
-        return rc
-
-    async def __call__(self, obj: filter.GetAttrObjectT) -> bool:
-        if self._used:
-            raise RuntimeError("RuleCondition cannot be reused")
-        self._used = True
-        try:
-            self.match = await self.partial_filter(obj)
-        except live_resolvers.LiveResolutionFailure as e:
-            self.match = False
-            self.evaluation_error = e.reason
-        return self.match
-
-    def get_attribute_name(self) -> str:
-        tree = typing.cast(filter.TreeT, self.partial_filter.tree)
-        tree = tree.get("-", tree)
-        name = list(tree.values())[0][0]
-        if name.startswith(filter.Filter.LENGTH_OPERATOR):
-            return str(name[1:])
-        return str(name)
-
-
-@dataclasses.dataclass
-class RuleConditionGroup:
-    """This describe a group leafs of the `conditions:` tree linked by and or or."""
-
-    condition: dataclasses.InitVar[
-        typing.Dict[
-            typing.Literal["and", "or"],
-            typing.List[typing.Union["RuleConditionGroup", RuleCondition]],
-        ]
-    ]
-    operator: typing.Literal["and", "or"] = dataclasses.field(init=False)
-    conditions: typing.List[
-        typing.Union["RuleConditionGroup", RuleCondition]
-    ] = dataclasses.field(init=False)
-    match: bool = dataclasses.field(init=False, default=False)
-    _used: bool = dataclasses.field(init=False, default=False)
-
-    def __post_init__(
-        self,
-        condition: typing.Dict[
-            typing.Literal["and", "or"],
-            typing.List[typing.Union["RuleConditionGroup", RuleCondition]],
-        ],
-    ) -> None:
-        if len(condition) != 1:
-            raise RuntimeError("Invalid condition")
-
-        self.operator, self.conditions = next(iter(condition.items()))
-
-    async def __call__(
-        self,
-        obj: filter.GetAttrObjectT,
-    ) -> bool:
-        if self._used:
-            raise RuntimeError("RuleConditionGroup cannot be re-used")
-        self._used = True
-        self.match = await filter.BinaryFilter(
-            typing.cast(
-                filter.TreeT,
-                {self.operator: self.conditions},
-            )
-        )(obj)
-        return self.match
-
-    def extract_raw_filter_tree(
-        self,
-        condition: typing.Optional[
-            typing.Union["RuleConditionGroup", RuleCondition]
-        ] = None,
-    ) -> filter.TreeT:
-        if condition is None:
-            condition = self
-
-        if isinstance(condition, RuleCondition):
-            return typing.cast(filter.TreeT, condition.partial_filter.tree)
-        elif isinstance(condition, RuleConditionGroup):
-            return typing.cast(
-                filter.TreeT,
-                {
-                    condition.operator: [
-                        self.extract_raw_filter_tree(c) for c in condition.conditions
-                    ]
-                },
-            )
-        else:
-            raise RuntimeError("unexpected condition instance")
-
-    def walk(
-        self,
-        conditions: typing.Optional[
-            typing.List[typing.Union["RuleConditionGroup", RuleCondition]]
-        ] = None,
-    ) -> typing.Iterator[RuleCondition]:
-        if conditions is None:
-            conditions = self.conditions
-        for condition in conditions:
-            if isinstance(condition, RuleCondition):
-                yield condition
-            elif isinstance(condition, RuleConditionGroup):
-                for _condition in self.walk(condition.conditions):
-                    yield _condition
-            else:
-                raise RuntimeError(f"Unsupported condition type: {type(condition)}")
-
-    def is_faulty(self) -> bool:
-        return any(c.evaluation_error for c in self.walk())
-
-    def copy(self) -> "RuleConditionGroup":
-        return RuleConditionGroup(
-            {self.operator: [c.copy() for c in self.conditions]},
-        )
-
-    @staticmethod
-    def _get_rule_condition_summary(cond: RuleCondition) -> str:
-        summary = ""
-        checked = "X" if cond.match else " "
-        summary += f"- [{checked}] `{cond}`"
-        if cond.description:
-            summary += f" [{cond.description}]"
-        if cond.evaluation_error:
-            summary += f" ⚠️ {cond.evaluation_error}"
-        summary += "\n"
-        return summary
-
-    @classmethod
-    def _walk_for_summary(
-        cls,
-        conditions: typing.List[typing.Union["RuleConditionGroup", RuleCondition]],
-        level: int = 0,
-    ) -> str:
-        summary = ""
-        for condition in conditions:
-            if isinstance(condition, RuleCondition):
-                summary += cls._get_rule_condition_summary(condition)
-            elif isinstance(condition, RuleConditionGroup):
-                label = "all of" if condition.operator == "and" else "any of"
-                checked = "X" if condition.match else " "
-                summary += f"- [{checked}] {label}:\n"
-                for _sum in cls._walk_for_summary(condition.conditions, level + 1):
-                    summary += _sum
-            else:
-                raise RuntimeError(f"Unsupported condition type: {type(condition)}")
-
-        return textwrap.indent(summary, "  " * level)
-
-    def get_summary(self) -> str:
-        return self._walk_for_summary(self.conditions)
-
-
-@dataclasses.dataclass
-class QueueRuleConditions:
-    conditions: dataclasses.InitVar[
-        typing.List[typing.Union["RuleConditionGroup", RuleCondition]],
-    ]
-    condition: RuleConditionGroup = dataclasses.field(init=False)
-    _evaluated_conditions: typing.List[
-        typing.Tuple[context.BasePullRequest, RuleConditionGroup]
-    ] = dataclasses.field(default_factory=list, init=False, repr=False)
-    match: bool = dataclasses.field(init=False, default=False)
-    _used: bool = dataclasses.field(init=False, default=False)
-
-    def __post_init__(
-        self,
-        conditions: typing.List[typing.Union["RuleConditionGroup", RuleCondition]],
-    ) -> None:
-        self.condition = RuleConditionGroup({"and": conditions})
-
-    def copy(self) -> "QueueRuleConditions":
-        return QueueRuleConditions(self.condition.copy().conditions)
-
-    def extract_raw_filter_tree(self) -> filter.TreeT:
-        return self.condition.extract_raw_filter_tree()
-
-    async def __call__(
-        self, pull_requests: typing.List[context.BasePullRequest]
-    ) -> bool:
-        if self._used:
-            raise RuntimeError("QueueRuleConditions cannot be re-used")
-        self._used = True
-
-        for pull in pull_requests:
-            c = self.condition.copy()
-            await c(pull)
-            self._evaluated_conditions.append((pull, c))
-
-        self.match = all(c.match for p, c in self._evaluated_conditions)
-        return self.match
-
-    def get_summary(self) -> str:
-        if self._used:
-            # TODO(sileht): Improve the summary for batch
-            return self._evaluated_conditions[0][1].get_summary()
-        else:
-            return self.condition.get_summary()
-
-    def is_faulty(self) -> bool:
-        if self._used:
-            return any(c.is_faulty() for p, c in self._evaluated_conditions)
-        else:
-            return self.condition.is_faulty()
-
-    def walk(self) -> typing.Iterator[RuleCondition]:
-        if self._used:
-            for _, conditions in self._evaluated_conditions:
-                for cond in conditions.walk():
-                    yield cond
-        else:
-            for cond in self.condition.walk():
-                yield cond
-
-
-async def get_branch_protection_conditions(
-    repository: context.Repository, ref: github_types.GitHubRefType
-) -> typing.List[typing.Union["RuleConditionGroup", RuleCondition]]:
-    return [
-        RuleCondition(
-            f"check-success-or-neutral={check}",
-            description="🛡 GitHub branch protection",
-        )
-        for check in await repository.get_branch_protection_checks(ref)
-    ]
-
-
-async def get_depends_on_conditions(
-    ctxt: context.Context,
-) -> typing.List[typing.Union["RuleConditionGroup", RuleCondition]]:
-    conds: typing.List[typing.Union["RuleConditionGroup", RuleCondition]] = []
-    for pull_request_number in ctxt.get_depends_on():
-        try:
-            dep_ctxt = await ctxt.repository.get_pull_request_context(
-                pull_request_number
-            )
-        except http.HTTPNotFound:
-            description = f"⛓️ ⚠️ *pull request not found* (#{pull_request_number})"
-        else:
-            description = f"⛓️ **{dep_ctxt.pull['title']}** ([#{pull_request_number}]({dep_ctxt.pull['html_url']}))"
-        conds.append(
-            RuleCondition(
-                {"=": ("depends-on", f"#{pull_request_number}")},
-                description=description,
-            )
-        )
-    return conds
-
-
 class DisabledDict(typing.TypedDict):
     reason: str
-
-
-@dataclasses.dataclass
-class PullRequestRuleConditions:
-    conditions: dataclasses.InitVar[
-        typing.List[typing.Union["RuleConditionGroup", RuleCondition]],
-    ]
-    condition: RuleConditionGroup = dataclasses.field(init=False)
-
-    def __post_init__(
-        self,
-        conditions: typing.List[typing.Union["RuleConditionGroup", RuleCondition]],
-    ) -> None:
-        self.condition = RuleConditionGroup({"and": conditions})
-
-    async def __call__(self, objs: typing.List[context.BasePullRequest]) -> bool:
-        if len(objs) > 1:
-            raise RuntimeError(
-                "PullRequestRuleConditions take only one pull request at a time"
-            )
-        return await self.condition(objs[0])
-
-    def extract_raw_filter_tree(self) -> filter.TreeT:
-        return self.condition.extract_raw_filter_tree()
-
-    def get_summary(self) -> str:
-        return self.condition.get_summary()
-
-    @property
-    def match(self) -> bool:
-        return self.condition.match
-
-    def is_faulty(self) -> bool:
-        return self.condition.is_faulty()
-
-    def walk(self) -> typing.Iterator[RuleCondition]:
-        for cond in self.condition.walk():
-            yield cond
-
-    def copy(self) -> "PullRequestRuleConditions":
-        return PullRequestRuleConditions(self.condition.copy().conditions)
 
 
 # TODO(sileht): rename me PullRequestRule ?
@@ -395,14 +43,14 @@ class PullRequestRuleConditions:
 class Rule:
     name: str
     disabled: typing.Union[DisabledDict, None]
-    conditions: PullRequestRuleConditions
+    conditions: conditions.PullRequestRuleConditions
     actions: typing.Dict[str, actions.Action]
     hidden: bool = False
 
     class T_from_dict_required(typing.TypedDict):
         name: str
         disabled: typing.Union[DisabledDict, None]
-        conditions: PullRequestRuleConditions
+        conditions: conditions.PullRequestRuleConditions
         actions: typing.Dict[str, actions.Action]
 
     class T_from_dict(T_from_dict_required, total=False):
@@ -430,12 +78,12 @@ QueueName = typing.NewType("QueueName", str)
 @dataclasses.dataclass
 class QueueRule:
     name: QueueName
-    conditions: QueueRuleConditions
+    conditions: conditions.QueueRuleConditions
     config: QueueConfig
 
     class T_from_dict(QueueConfig, total=False):
         name: QueueName
-        conditions: QueueRuleConditions
+        conditions: conditions.QueueRuleConditions
 
     @classmethod
     def from_dict(cls, d: T_from_dict) -> "QueueRule":
@@ -449,12 +97,12 @@ class QueueRule:
         ref: github_types.GitHubRefType,
         pulls: typing.List[context.BasePullRequest],
     ) -> EvaluatedQueueRule:
-        branch_protection_conditions = await get_branch_protection_conditions(
-            repository, ref
+        branch_protection_conditions = (
+            await conditions.get_branch_protection_conditions(repository, ref)
         )
         queue_rule_with_branch_protection = QueueRule(
             self.name,
-            QueueRuleConditions(
+            conditions.QueueRuleConditions(
                 self.conditions.condition.copy().conditions
                 + branch_protection_conditions
             ),
@@ -617,16 +265,18 @@ class PullRequestRules:
                     actions_without_special_rules[name] = action
 
             if actions_with_special_rules:
-                branch_protection_conditions = await get_branch_protection_conditions(
-                    ctxt.repository, ctxt.pull["base"]["ref"]
+                branch_protection_conditions = (
+                    await conditions.get_branch_protection_conditions(
+                        ctxt.repository, ctxt.pull["base"]["ref"]
+                    )
                 )
-                depends_on_conditions = await get_depends_on_conditions(ctxt)
+                depends_on_conditions = await conditions.get_depends_on_conditions(ctxt)
                 if branch_protection_conditions or depends_on_conditions:
                     runtime_rules.append(
                         Rule(
                             name=rule.name,
                             disabled=rule.disabled,
-                            conditions=PullRequestRuleConditions(
+                            conditions=conditions.PullRequestRuleConditions(
                                 rule.conditions.condition.copy().conditions
                                 + branch_protection_conditions
                                 + depends_on_conditions
@@ -733,7 +383,7 @@ def RuleConditionSchema(v: typing.Any, depth: int = 0) -> typing.Any:
 
     return voluptuous.Schema(
         voluptuous.Any(
-            voluptuous.All(str, voluptuous.Coerce(RuleCondition)),
+            voluptuous.All(str, voluptuous.Coerce(conditions.RuleCondition)),
             voluptuous.All(
                 {
                     "and": voluptuous.All(
@@ -746,7 +396,7 @@ def RuleConditionSchema(v: typing.Any, depth: int = 0) -> typing.Any:
                     ),
                 },
                 voluptuous.Length(min=1, max=1),
-                voluptuous.Coerce(RuleConditionGroup),
+                voluptuous.Coerce(conditions.RuleConditionGroup),
             ),
         )
     )(v)
@@ -764,7 +414,7 @@ def get_pull_request_rules_schema(partial_validation: bool = False) -> voluptuou
                     voluptuous.Required("hidden", default=False): bool,
                     voluptuous.Required("conditions"): voluptuous.All(
                         [voluptuous.Coerce(RuleConditionSchema)],
-                        voluptuous.Coerce(PullRequestRuleConditions),
+                        voluptuous.Coerce(conditions.PullRequestRuleConditions),
                     ),
                     voluptuous.Required("actions"): actions.get_action_schemas(
                         partial_validation
@@ -784,7 +434,7 @@ QueueRulesSchema = voluptuous.All(
                 voluptuous.Required("name"): str,
                 voluptuous.Required("conditions"): voluptuous.All(
                     [voluptuous.Coerce(RuleConditionSchema)],
-                    voluptuous.Coerce(QueueRuleConditions),
+                    voluptuous.Coerce(conditions.QueueRuleConditions),
                 ),
                 voluptuous.Required("speculative_checks", default=1): voluptuous.All(
                     int, voluptuous.Range(min=1, max=20)
